Suppress expected output from passing tests
Move scoped_test_stderr into a lightweight shared test header so core
and plugin tests can use it without importing task-code definitions.
Expected diagnostics are discarded for successful tests and replayed
when GoogleTest records a failure.

Apply the helper to the simple logger test and update the service
submodule to merged PR #36, which suppresses expected service-test and
simple-KV traces in the same success-only manner.
